update(uploadFilesToConvert): rename file and restructure formdata
- Rename file to: uploadFileToConvert.
- Restructure formdata to append a single file with it's corresponding file_config.
